+++ Logs/Changelog.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version [2.0.0] - 2024-03-19
+
+## Added
+- Functionality to Save Favorites
+- Saving Logs for Favorites updation and Errors
+- Videos Count in Summary Report
+- File Loading by Saving Hashed Filenames Csv And Storing Media Files Locations For Fast Access
+- Getting Files and Csvs Together in a Single Input
+- Feedback label To Display Videoname and display Favorites Trigger
+- Full Screen Functionality
+- Saving Screen Capture of any Playing video.
+
+## Changed
+- Player Window Size from **800x600** to **1000x600**
+- Media Playing by Skiping any File That Doesn't Exist
+- Length of Media Progress Bar from **600** to **800**
+
 ## Version [1.1.0] - 2024-03-14
 
 ### Added